Use map reduce pattern for 2i
This feels natural to do in python and I wanted to see what it would
look like in rust. Frankly I do not think it is any easier to read and
I expect it would be harder if the reader is not familiar with the
map reduce function from more itertools in Python. But the results are
nonetheless interesting.