Add solution for d25
Porting this from python was pretty straight forward except for the
loop; I tried for a long time to implement something like

  new, old = func(new), new

but it would either complain about the value being moved or not update
new properly causing each iteration of the loop to be like the first.
The solution that worked in the end is hard to read in my opinion
because keeping track of which side of the array we should access is
not intuitive for me. I still hope that what I wanted to do first is
possible but I am tired of searching for the answer for now.