Add solution for d23
Initially ported straight from python but performance was miserable,
only 50% faster with the release flag. In hindsight I realize that
was probably due to me caching incorrectly. I tried briefly to profile
the program and it seemed to show that a lot of time was spent
calculating the hashes so I blamed that, hence the u64 representation
of hallway and rooms.

While I am not happy with the verbose and dangerous code it is 3x faster
than using Vectors for hallway and rooms.

Test duration (signle threaded): 0.11s