wip: replace open set with cost comparison
Inspired by reading other implementations of a star.

This takes me closer to my goal of eliminating the clones (would be
nice to 1. not have to implement the trait and 2. avoid the overhead).

I expect the speedup is mostly thanks to not having to compute hashes
and (re)allocating the hashset.

Before: 0.20
After: 0.18