Remove iter_zip feature
This removes the dependency on nightly which, with a bit of luck, will
help making the crate installable from crates.io. So far it has been
complaining about the below but with a stable version of cargo at least
it downloads the crate and complains that it cannot be built.

  could not find `aocoracle` in registry `crates-io` with version `*`

Also sneak in a clippy fix.