Add github action running various checks
This should help ensure that main is always a good base of which to
start new changes.

Stops short of publishing the crate because I need to figure out how it
should be triggered. In python I would trigger it when a tag is pushed
thanks to setuptools-scm using the git tag as package version but I do
not know of an equivalent function in rust (and I am not entirely sure
I like that workflow anyway since the tag is not an immutable part of
git history).