Add manual workflow to publish crate
By publishing via a workflow I make it harder for myself to publish a
commit that does not pass all tests or that is not part of of the git
history. I say harder because there are several shortcomings, most
importantly that the workflow can be run for any branch and only the
main branch is protected.

Remove use of secrets from the integrate workflow because they are not
needed and they would not be available for pull requests from forks.