Create simple webapp showcasing the program
Put the webapp in its own crate to increase separation and because I
anticipate adding more crates such as maybe a proc macro.

Call the crates directory crates because it mirrors the name packages
that I am used to from python workspace repos and it is the name used
by cargo.

Move aoclib into crates for consistency.