Replace test detecting false negatives
For this it is useful to have a central repository of available
input-answer pairs and when we have that we might as well use that in
the individual modules. This makes the module tests somewhat redundant
(the error of one part could theoretically be offset by an incorrect
answer of another, but this seems unlikely) but it is convenient to get
a test report for each and to have a test to click on in the IDE to run
it. It is also immediately clear what part of the code is failing.