Add early exit as default behavior
The goal with the program is less to be useful and more to be fun and
to me it seems more fun if the program returns exactly one answer or
no answer at all.

The --exhaustive flag can be set to tell the program to present all
feasible answers. In this case it will also say what day it is from
because it is interesting to see what days are easily confused.