Optimize y2018d03
Can probably do something with better performance by utilizing quad-
trees or splitting the rectangles somehow but this is good enough for
me at the moment.

Also add make shortcut for measuring duration of tests as a quick way
to profile what solvers should be improved and measure improvements.
It is not obvious to me that I want this to be part of check_all; on
the one hand it is not guaranteed to work because I don't want to spend
CI time on this test. On the other I want to remember to run it locally

This change decreases the duration of part_1_works_on_input from 0.05s
to 0.02s.

The duration of returns_error_on_wrong_input could be improved from
0.10s to 0.02s by compiling the regex only once. But since it is only
compiled once during normal operation eagerly compiling it would slow
down the program since it would compile even when it does not run. It
is may be possible to get the same effect by using a static variable in
the function but then we quickly run into issues with thread safety.