y2018d02: Encode in type that ids are byte strings
I don't think the solutions ever assumed that this is the case but the
parsing already ensured it so I thought it would be fun to try. I think
the code is slightly uglier for it and the crude benchmarks I have
cannot demonstrate that it changes the performance, for better or
worse. But I keep it in anyway as a reminder of how it looked with and
without.

I tried to make the type &[u8] because it looks nicer but I could not
quickly figure out how to make the lifetimes work out.