Allow user to cycle through answers
For some inputs it will be impossible to tell what problem they
correspond to and it would be nice to still have a way to show that
the app finds the correct answer.

As of writing this is not a problem because while there are inputs
that will be answered by the wrong solver no two solvers are able to
answer the same input.