Factor out rectangle
... to explore limitations and workarounds for using generic types.
Making it generic seems like a good way to remove some code duplication
within a crate since one can always fall back on implementing methods
for concrete types.

Prevent direct instantiation to guarantee that width and height are
non-negative.

Rename overlap -> intersection since the latter is more common and
with the type factored out it is more important that names make sense
in the context of the type as opposed to the specific problem where the
type is employed.

Rename coordinates -> tiles to reflect that they are discrete and their
numbers therefore finite.