Use macro for standard tests
Also update all tests to use the same functions to make it easier to
read a second module once one has read the first.

Keep most code outside of macros to
* improve support from tools including tab completion and formatting,
* ensure that panics from unwrap originate on a line of real code that
  can be navigated to.

Rename _ functions to reveal dead code. These were originally named
that way because in python this is a convenient way to separate
protected functions and avoid name collisions.

Change the generics of actual- and expected-answer functions so that
the same function can be used in more places hoping that this will
avoid confusion around why there are two functions and what the
difference is.

Increase the number of expected true positives since there are now more
examples among the answers.