y2022d01: Factor out top 3 collection
... to make the implementation read more naturally.

Also change the implementation to use a BinaryHeap instead of sorting
all items because the asymptotic runtime is better. Probably does not
help much, if at all, in practice.

It annoys me that part_2 has to call into_iter on the resulting vector
but I could not find a neat way to return an iterator.