Add option to save session cookies
Also make it possible to override the user fingerprint value so that
only one API call needs to be made when retrieving the input in the
morning.

Also separate input and answer scraping since the answer scraping
cannot be cached, building the cache takes many requests and I don't
want to spam the API unnecessarily.

Multiple session cookies are supported to make it easier expanding the
test data using multiple logins.

Also format the answers as they will be formatted by the scraper to
avoid diffs in the future when the scraper is used.