Reorganize data as files named from input content
Storing them as individual files will avoid contention and merge
conflicts in answers.json.

Naming them after the content will avoid bad conflicts of fingerprints;
since every user does not get a unique input the fingerprint will not
be unique either. Now, if two users who share input for the first
problem don't share input for another problem the system with they will
disagree on what the content of that file should be.

To accommodate the examples and other non-official inputs without
creating an even deeper hierarchy the convention that official inputs
have lower case names and other inputs have upper case names is
established. This only works on case sensitive file systems, but the
risk of collisions seems low enough that it should be viable
everywhere.

Because naming files after the content makes the name unpredictable
today's inputs and answers are also aliased as INPUT. When preparing a
new module for the next days problem this name can be used to save some
typing.

Because some inputs were the same for multiple users the total number
of test inputs is lower after this change.