Add solutions for y2023d04 (#4)
Numbers are no more than two digits so parsing them as u8 will reject
more inputs and, maybe, be faster.

Card numbers are consecutive so we can use a vector instead of a map.
Together with changing the type of the numbers to u8, this reduces the
time it takes to run part 2 from about 18ms to 11ms.

Parse sequences of numbers using `flat_map` because `?` does not work
from inside `map`. This could make some bugs hard to find, but it
works for now. I'd like to find a variant of map that returns `Err` if
any element was `Err` and `Ok` if all elements were `Ok`.

It is possible to solve pt2 naively and the solution runs in about 1s.
But by using a cache the time is reduced to about 10ms.